Rewrite, don't use the token_serie and store hashed token in database.

+++ auth_remember/backend.py
@@ -1,6 +1,9 @@
+from datetime import datetime, timedelta
+
 from django.contrib.auth.models import User
 
 from auth_remember import auth
+from auth_remember import settings
 from auth_remember.models import RememberToken
 
 
@@ -10,37 +13,26 @@ class RememberBackend(object):
 
     """
     def authenticate(self, remember_token, request):
-        """Return the user associated with the given token.
-
-        If the token is invalid (no valid user found) then delete all
-        tokens associated with the token serie.
-
-        """
-        try:
-            serie_token, user_id, token = remember_token.split(':')
-        except ValueError:
-            auth.preset_cookie(request, '')
+        """Return the user associated with the given token."""
+        token = RememberToken.objects.get_by_string(
+            token_string=remember_token)
+        if not token:
             return
 
-        try:
-            token = RememberToken.objects.select_related('user').get(
-                token=token,
-                serie_token=serie_token,
-                user=user_id)
-
-            # Create new token cookie value
-            new_token = auth.create_token_object(token.user, token)
+        # If the token is older then COOKIE_AGE then delete it and return
+        max_age = datetime.now() - timedelta(seconds=settings.COOKIE_AGE)
+        if token.created_initial < max_age:
             token.delete()
-            auth.preset_cookie(request, new_token)
+            return
 
-            return token.user
-        except RememberToken.DoesNotExist:
+        user = token.user
 
-            # We might be dealing with a stolen cookie, delete all the
-            # remember me information from the user.
-            RememberToken.objects.filter(
-                user=user_id, serie_token=serie_token).delete()
-            auth.preset_cookie(request, '')
+        # Create new token cookie value and delete current token
+        token_string = auth.create_token_string(user, token)
+        auth.preset_cookie(request, token_string)
+        token.delete()
+
+        return user
 
     def get_user(self, user_id):
         try:
